Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.5.3
